Move STT hook to hooks directory
